write tests for TestMedicine, tighten CRUD to throw exceptions in genericRepo

+++ src/repository/GenericRepository.java
@@ -47,23 +47,27 @@ public class GenericRepository<T extends IdentifiedObject> extends AbstractRepos
     }
 
     protected void defaultCreateItem(T item) {
+        if (db.containsKey(item.getID())) throw new RuntimeException("Create operation failed! Item does not exist!");
         db.put(item.getID(), item);
         saveDatabase();
     }
 
     protected T defaultReadItem(String item_id){
+        if (!db.containsKey(item_id)) throw new RuntimeException("Read operation failed! Item does not exist!");
         return db.get(item_id);
     }
 
 
     protected void defaultUpdateItem(T item){
         String item_id = item.getID();
+        if (!db.containsKey(item_id)) throw new RuntimeException("Update operation failed! Item does not exist!");
         db.remove(item_id); // remove old item
         db.put(item_id, item); // add new item
         saveDatabase();
     }
 
     protected void defaultDeleteItem(String item_id){
+        if (!db.containsKey(item_id)) throw new RuntimeException("Delete operation failed! Item does not exist!");
         db.remove(item_id);
         saveDatabase();
     }
